requirements.txt
# Core dependencies
numpy>=1.20.0
pandas>=1.3.0
scikit-learn>=1.0.0
matplotlib>=3.4.0
pyyaml>=6.0


# Web scraping and processing
requests>=2.26.0
beautifulsoup4>=4.10.0
lxml>=4.6.3
crawl4ai>=0.4.1

# Topic modeling
bertopic>=0.14.0
sentence-transformers>=2.2.0
hdbscan>=0.8.28
umap-learn>=0.5.2

# Async processing
aiohttp>=3.8.1
asyncio>=3.4.3

# LLM integration
together>=0.1.5
google-generativeai>=0.3.1  # Google Gemini API

# RAG pipeline dependencies
chromadb>=0.4.6  # Vector database (local option)
#pinecone-client>=2.2.1  # Vector database (cloud option)
#qdrant-client>=1.1.1  # Alternative vector database
faiss-cpu>=1.7.3  # Facebook AI Similarity Search
regex>=2022.3.15  # Enhanced regular expressions
argparse>=1.4.0  # Command-line argument parsing

# Utilities
tqdm>=4.62.3
python-dotenv>=1.0.0  # For managing environment variables

# Reddit API
praw>=7.0.0



